backend/requirements.txt
asgiref==3.5.2
Brotli==1.0.9
Django==4.1
django-cors-headers==3.13.0
django-crontab==0.7.1
django-dbbackup==4.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
whitenoise==6.4.0
